requirements.txt
fastapi
uvicorn
jinja2
python-multipart
aiohttp
aiomysql
itsdangerous
python-dotenv
